requirements.txt
aiogram>=3,<4
fastapi>=0.111
uvicorn[standard]>=0.29
python-dotenv>=1.0
